+++ src/pages/Calendar/Calendar.jsx
@@ -1,7 +1,7 @@
-import { useState } from "react";
-import "./Calendar.css";
 import getDayDiff from "../../helpers/getDayDiff";
 import CalendarDay from "./ui/CalendarDay";
+import { useState, useEffect } from "react";
+import "./Calendar.css";
 
 export default function Calendar(){
     if(localStorage.getItem("calendar-index") === null){
@@ -9,12 +9,19 @@ export default function Calendar(){
     }
     const [isRender, setIsRender] = useState(false);
     let localKeys = Object.keys(localStorage);
+
+    const [activeId, setActiveId] = useState(null)
+
+    const [draggedKey, setDraggedKey] = useState(null)
+    const [selectedDay, setSelectedDay] = useState(null)
+    const [pos, setPos] = useState({ x: 0, y: 0 }) // може у localStorage?? щоб можна було звертатись коли захочеш і без зайвих рендерів
+    // console.log(draggedKey)
     
     let calendar = {};
     for(let i = 0; i < localKeys.length; i++){ // Створюємо об'єкт в якому будуть наші ключики і значення
         if(localKeys[i].includes("calendar-item")){
             let localArr = localStorage.getItem(localKeys[i]).split("^")
-            console.log(localArr);
+            // console.log(localArr);
             let date = localArr[2]+"." + localArr[3]+"." + localArr[4];
 
             let newDateFormat = date.split(".");
@@ -23,7 +30,7 @@ export default function Calendar(){
             newDateFormat[2] = temp;
             let newDate = newDateFormat.map(el => el.padStart(2, "0")).join(".");
             let dayDiff = getDayDiff(newDate);
-            console.log(dayDiff);
+            // console.log(dayDiff);
             if(dayDiff >= 1) {
                 if(!calendar.expired){
                     calendar.expired = [localKeys[i]]
@@ -43,14 +50,16 @@ export default function Calendar(){
         date.setDate(date.getDate() + i-1);
         sevenDays.push(`${date.getFullYear()}.${date.getMonth() + 1}.${date.getDate()}`);
     }
-    console.log(sevenDays);
+    // console.log(sevenDays);
     for(let i = 0; i < sevenDays.length; i++){
         if(calendar[sevenDays[i]] === undefined) {
             calendar[sevenDays[i]] = [];
         }
     }
-    console.log(calendar);
-    console.log(Object.keys(calendar));
+    for(let key in calendar){ // сортуємо задачі (спочатку найновіші потім внизу найстаріші)
+        calendar[key].sort((a,b) => +a.split("-")[2] - +b.split("-")[2])
+    }
+    // console.log(Object.keys(calendar));
     let calendarKeys = Object.keys(calendar);
     calendarKeys.sort((a, b) => {
         if(a === "expired") return -1;
@@ -59,14 +68,24 @@ export default function Calendar(){
         let dateB = new Date(...b.split(".").map(Number));
         return dateA - dateB;
     });
+
+    useEffect(() => {
+        const handleMouseMove = (event) => {
+            setPos({ x: event.clientX, y: event.clientY });
+        };
+        window.addEventListener("mousemove", handleMouseMove);
+        return () => {
+            window.removeEventListener("mousemove", handleMouseMove);
+        };
+    }, [])
+
     return (
         <div className="calendar content">
             <div className="calendar__content newblock">
-                {calendarKeys.map(el => {
-                    return <CalendarDay key={el} date={el} keyArr={calendar[el]} onChange={() => setIsRender(!isRender)} />;
+                {calendarKeys.map((el, index) => {
+                    return <CalendarDay key={el} date={el} keyArr={calendar[el]} onChange={() => setIsRender(!isRender)} activeId={activeId} setActiveId={setActiveId} index={index} draggedKey={draggedKey} setDraggedKey={setDraggedKey} pos={pos} setPos={setPos} selectedDay={selectedDay} setSelectedDay={setSelectedDay}/>;
                 })}
             </div>
         </div>
     );    
 }
-// ЗРОБИ ТАК ЩОБ НЕ ВИВОДИЛОСЬ ПОЗАПОЗАВЧОРА, ТІЛЬКИ ЩОБ ВЧОРА І ПОЗАВЧОРА